regcomp.c - debug output changes to be less cluttered
The output of use 're' debug was getting very difficult to
understand. This patches declutters the output by removing
most redundant data about the next regop that would be processed
following a given regop. Only when the "next" regop data would
*differ* from the "following regnode" is it shown. When it /is/
shown it is not shown in parenthesis but with an arrow notation.

Previous to this patch we might get output like this:
    Compiling REx "foo(?:[Bb]ar|fop)*baz"
    Final program:
       1: EXACT <foo> (3)
       3: CURLYM[0]{0,INFTY} (16)
       5:   BRANCH (buf:0) (10)
       6:     ANYOFM[Bb] (8)
       8:     EXACT <ar> (14)
      10:   BRANCH (buf:0) (FAIL)
      11:     EXACT <fop> (14)
      13:   TAIL (14)
      14:   SUCCEED (0)
      15: NOTHING (16)
      16: EXACT <baz> (18)
      18: END (0)

After this patch we get more readable (IMO) output like this:
    Compiling REx "foo(?:[Bb]ar|fop)*baz"
    Final program:
       1: EXACT <foo>
       3: CURLYM[0]{0,INFTY} -> 18
       5:   BRANCH (buf:0) -> 11
       7:     ANYOFM[Bb]
       9:     EXACT <ar> -> 16
      11:   BRANCH (buf:0) -> FAIL
      13:     EXACT <fop> -> 16
      15:   TAIL
      16:   SUCCEED.
      17: NOTHING
      18: EXACT <baz>
      20: END.

Transitions to 0, which really mean that the regop is an "END" type
regop are indicated by a full stop. Non linear transitions or
transitions to a regop that is not the physical successor to the
"current" regnode are shown as -> %d, or -> FAIL for fail case branches.

Notice the difference between

	5:   BRANCH (buf:0) (10)
       	5:   BRANCH (buf:0) -> 11